fix: complete governed extension integration

+++ backend/app/workforce/extensions/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, JSON
+from sqlalchemy import BigInteger, Column, Integer, String, JSON
 from sqlalchemy.schema import UniqueConstraint
 from app.db.base_class import Base
 
@@ -6,7 +6,7 @@ class ExtensionRegistration(Base):
     __tablename__ = "extension_registrations"
 
     id = Column(Integer, primary_key=True, index=True)
-    workspace_id = Column(Integer, nullable=False)
+    workspace_id = Column(BigInteger, nullable=False)
     extension_id = Column(String, nullable=False)
     version = Column(String, nullable=False)
     status = Column(String, nullable=False)
